# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fission HTTP Server API
  description: |
    REST API for managing Fission resources on Kubernetes.
    This API wraps all fission-cli commands and provides HTTP endpoints for:
    - Functions (create, list, get, update, delete, test, logs, pods, metadata)
    - Environments (create, list, get, update, delete, pods)
    - Packages (create, list, get source/deploy/info, update, delete, rebuild)
    - HTTP Triggers (create, list, get, update, delete)
    - Time Triggers (create, list, delete)
    - MQ Triggers (create, list, delete)
    - KubeWatches (create, list, delete)
    - Version information
    - System check
    - Plugins
  version: 2.0.0
  contact:
    name: Fission HTTP Server
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server (example)
paths:
...
  /api/v1/timetriggers:
    get:
      summary: List all time triggers
      description: Lists all time triggers.
      operationId: listTimeTriggers
      tags:
        - Time Triggers
      parameters:
        - name: namespace
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a time trigger
      description: Creates a new time trigger (cron job).
      operationId: createTimeTrigger
      tags:
        - Time Triggers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                function:
                  type: string
                cron:
                  type: string
      responses:
        '201':
          description: Time trigger created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/timetriggers/{name}:
    delete:
      summary: Delete a time trigger
      description: Deletes a time trigger.
      operationId: deleteTimeTrigger
      tags:
        - Time Triggers
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Time trigger deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  name:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # MQ Trigger endpoints
  /api/v1/mqtriggers:
    get:
      summary: List all MQ triggers
      description: Lists all message queue triggers.
      operationId: listMQTriggers
      tags:
        - MQ Triggers
      parameters:
        - name: namespace
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create an MQ trigger
      description: Creates a new message queue trigger.
      operationId: createMQTrigger
      tags:
        - MQ Triggers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                function:
                  type: string
                mqtype:
                  type: string
                topic:
                  type: string
      responses:
        '201':
          description: MQ trigger created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/mqtriggers/{name}:
    delete:
      summary: Delete an MQ trigger
      description: Deletes a message queue trigger.
      operationId: deleteMQTrigger
      tags:
        - MQ Triggers
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: MQ trigger deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  name:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # KubeWatch endpoints
  /api/v1/kubewatches:
    get:
      summary: List all KubeWatches
      description: Lists all Kubernetes watches.
      operationId: listKubeWatches
      tags:
        - KubeWatches
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a KubeWatch
      description: Creates a new Kubernetes watch trigger.
      operationId: createKubeWatch
      tags:
        - KubeWatches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                function:
                  type: string
                type:
                  type: string
                  description: Kubernetes object type to watch
      responses:
        '201':
          description: KubeWatch created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/kubewatches/{name}:
    delete:
      summary: Delete a KubeWatch
      description: Deletes a Kubernetes watch trigger.
      operationId: deleteKubeWatch
      tags:
        - KubeWatches
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: KubeWatch deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  name:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # System endpoints
...
  /api/v1/plugins:
    get:
      summary: List plugins
      description: Lists available Fission plugins.
      operationId: listPlugins
      tags:
        - System
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error message
          example: "An error occurred"
      required:
        - error